Solves yet another issue with BreakIterator.
The problem we're solving here is that sometimes we'll see text (e.g.,
from Microsoft Word) with a "right single quotation mark" (U+2019) as
an apostrophe, which java.text.BreakIterator will see as a word
break. So a string like "we’ll" gets tokenized as "we" and "ll", and
then "ll" is in no word list, and is flagged as incorrect.

The solution here is to simply find and replace that
character. Consequently, we now keep a copy of the "raw" text supplied
to the Tokenizer in the case that it was modified.